requirements.txt
streamlit>=1.32.0
anthropic>=0.40.0
pandas>=2.0.0
openpyxl>=3.1.0
pdfplumber>=0.10.0
beautifulsoup4>=4.12.0
requests>=2.31.0
plotly>=5.18.0
python-dotenv>=1.0.0
